AlertingSettingsIT created
